Unset new AM_CONDITIONAL for media direct on ia64
Add a new AM_CONDITIONAL BUILD_MEDIA_DIRECT which is set to false on
ia64.  The manual assembly in the media direct sources fails to build on
ia64 as it relies on more registers than available on ia64; there's no
ia64 system with media direct, so it should be disabled on ia64.