Fix to address UnicodeDecodeError Bug in smbios-token-ctl. Patch 1/3 Bug BZ1235061 raised by Scott Williams(vwfoxguru@gmail.com)in Fedora22.
Patch is to fix the UnicodeDecodeError which is mainly because of
few special characters present in place of apostrophes in
token_list.csv file. Cleaning up token_list.csv file by replacing
those characters with the proper apostrophes.

Tried differnt ways to address the encoding, couldnt find one that
works across differnt distros. Decided to clean up the token_list.csv
file instead.

PS - This bug is not seen on Ubuntu, encoding succeeds in that!!!

Signed-off-by: SriniG <srinivas_g_gowda@dell.com>