Fix a wrong conditional.
Covscan noticed:
Error: CONSTANT_EXPRESSION_RESULT (CWE-398): [#def22]
libsmbios-2.3.3/src/libsmbios_c/smbios/smbios_obj.c:442: always_true_or: The "or" condition "tempTEP->major_ver != 2 || tempTEP->major_ver != 3" will always be true because "tempTEP->major_ver" cannot be equal to two different values at the same time, so it must be not equal to at least one of them.

This code cannot possibly be right, so I'm assuming the idea is to fail
if it's not either 2 or 3.

This patch changes it to check that.

Signed-off-by: Peter Jones <pjones@redhat.com>