Fix some impossible logic.
Covscan says:
Error: CONSTANT_EXPRESSION_RESULT (CWE-398): [#def37]
libsmbios-2.3.3/src/libsmbios_c/smi/smi_password.c:201: impossible_and: The "and" condition "tmpret == 0 && tmpret == 2" can never be true because "tmpret" cannot be equal to two different values at the same time.

And it's right, that can't be correct.

I've re-written most of the function here - I think it's "correct", but
the behavior doesn't appear to be well defined, so it's hard to be sure.

Signed-off-by: Peter Jones <pjones@redhat.com>