Add "Proper" support for SMBIOS 3.0 tables
The existing claim of matching new SMBIOS table versions was a flat
out lie.  The support wasn't there and we were just lucky that
things worked on a variety of systems because the DMI entry point
was found by parsing /dev/mem.

Instead parse from the sysfs entry point and DMI table like apps
such as dmidecode do and add in checks for SMBIOS2.0 and SMBIOS3.0.